move ComponentWaitService & small convention fixes
