+++ @bellatrix/web/src/services/index.ts
@@ -3,6 +3,7 @@ export * from './BrowserAutomationToolLaunchService';
 export * from './ComponentService';
 export * from './NavigationService';
 export * from './CookiesService';
+export * from './ComponentWaitService';
 export * from './BrowserService';
 export * from './ScriptService';
 export * from './DialogService';
